+++ gazebo/gui/ModelAlign.cc
@@ -46,9 +46,19 @@ ModelAlign::ModelAlign()
 /////////////////////////////////////////////////
 ModelAlign::~ModelAlign()
 {
-  this->dataPtr->modelPub.reset();
+  this->Clear();
   delete this->dataPtr;
   this->dataPtr = NULL;
+}
+
+/////////////////////////////////////////////////
+void ModelAlign::Clear()
+{
+  this->dataPtr->userCamera.reset();
+  this->dataPtr->scene.reset();
+  this->dataPtr->node.reset();
+  this->dataPtr->modelPub.reset();
+  this->dataPtr->initialized = false;
 }
 
 /////////////////////////////////////////////////